OPS: Replace update-version with check-semantic-version job
