+++ .github/workflows/python-ci.yml
@@ -10,14 +10,17 @@ on: [push]
 
 jobs:
 
-  check-version-consistency:
+  check-semantic-version:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
-      - run: python .github/workflows/scripts/check-version-consistency.py
+      - name: Install git-mkver
+        run: curl -L https://github.com/idc101/git-mkver/releases/download/v1.2.1/git-mkver-linux-amd64-1.2.1.tar.gz | tar xvz && sudo mv git-mkver /usr/local/bin
+      - name: Check version
+        run: python .github/workflows/scripts/check-semantic-version.py
 
-  tests:
+  run-tests:
     if: "!contains(github.event.head_commit.message, 'skip_ci_tests')"
     runs-on: ${{ matrix.os }}
     env:
@@ -77,4 +80,3 @@ jobs:
 #          user: __token__
 #          password: ${{ secrets.PYPI_TOKEN }}
 #          verbose: true
-#
