+++ .github/workflows/release.yml
@@ -8,28 +8,6 @@ on:
       - main
 
 jobs:
-  update-version:
-    if: github.event.pull_request.merged == true
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-      with:
-        ref: main
-    - name: Install git-mkver
-      run: curl -L https://github.com/idc101/git-mkver/releases/download/v1.2.1/git-mkver-linux-amd64-1.2.1.tar.gz | tar xvz && sudo mv git-mkver /usr/local/bin
-    - name: Update version in setup.py
-      run: git-mkver patch
-    - name: Get new version
-      run: echo "NEW_PACKAGE_VERSION=$( git-mkver next )" >> $GITHUB_ENV
-    - name: Commit and push change
-      uses: EndBug/add-and-commit@v7
-      with:
-        add: "setup.py"
-        message: "OPS: Update version number"
-        branch: main
-        push: "true"
-        tag: ${{ env.NEW_PACKAGE_VERSION }}
-
 #  release:
 #    # This job will only run if the PR has been merged (and not closed without merging).
 #    if: github.event.pull_request.merged == true
